+++ src/app/chat/chat.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import { User } from '../interfaces/user';
+import { UserService } from '../services/user.service';
 
 @Component({
   selector: 'app-chat',
@@ -8,9 +10,17 @@ import { ActivatedRoute } from '@angular/router';
 })
 export class ChatComponent implements OnInit {
   friendId: any;
-  constructor(private activatedRoute: ActivatedRoute) { 
+  friends: User[];
+  friend: User;
+  constructor(private activatedRoute: ActivatedRoute,
+              private userService: UserService) { 
     this.friendId = this.activatedRoute.snapshot.params['uid'];
-    console.log(this.friendId)
+    this.friends = this.userService.getFriends();
+    
+    this.friend = this.friends.find(record =>{
+      return record.uid == this.friendId;
+    });
+    console.log(this.friend)
   }
 
   ngOnInit() {
