+++ src/app/home/home.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { User } from '../interfaces/user';
-
-
+import { UserService } from '../services/user.service';
 
 @Component({
   selector: 'app-home',
@@ -9,50 +8,10 @@ import { User } from '../interfaces/user';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
-
-    friends: User[];
-
-  constructor() {
-    let usuario1: User = {
-      nick: "Mike",
-      subnick: "hola",
-      age: 27,
-      email: "jose@gmail.com",
-      friend: true,
-      uid: 1
-    };
-    let usuario2: User = {
-      nick: 'Eduardo',
-      age: 24,
-      email: 'ed@aoe.aoe',
-      friend: true,
-      uid: 2
-    };
-    let usuario3: User = {
-      nick: 'Freddy',
-      age: 28,
-      email: 'fred@aoe.aoe',
-      friend: true,
-      uid: 3
-    };
-    let usuario4: User = {
-      nick: 'Yuliana',
-      age: 18,
-      email: 'yuli@aoe.aoe',
-      friend: true,
-      uid: 4
-    };
-    let usuario5: User = {
-      nick: 'Ricardo',
-      age: 17,
-      email: 'rick@aoe.aoe',
-      friend: false,
-      uid: 5
-    };
-    
-    this.friends = [usuario1, usuario2, usuario3, usuario4, usuario5];
-   
-   }
+  friends: User[];
+  constructor(private userService: UserService) {
+    this.friends = userService.getFriends();
+  }
 
   ngOnInit() {
   }
